Get rid of ChildrenContext
